src/test/java/org/alan/algorithm/practice/baekjoon/clazz/four/BJ1915Test.java
package org.alan.algorithm.practice.baekjoon.clazz.four;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

public class BJ1915Test {
    static Stream<Arguments> parameters() {
        return Stream.of(
                Arguments.of("""
                        4 4
                        0100
                        0111
                        1110
                        0010""", "4")
        );
    }

    @ParameterizedTest
    @MethodSource("parameters")
    public void test1(String input, String expected) {
        Assertions.assertEquals(expected, new BJ1915().solve(input));
    }
}
